23-085 tugas5done

+++ Modul 5/23-085_Bangkit_Rizky_Lillah/hapus.php
@@ -2,7 +2,17 @@
 include 'koneksi.php';
 $id = $_GET['id'] ?? 0;
 
-mysqli_query($koneksi, "DELETE FROM supplier WHERE id=$id");
-header("Location: index.php");
-exit;
+// Cek apakah supplier masih digunakan
+$cek = mysqli_query($koneksi, "SELECT COUNT(*) AS jml FROM barang WHERE supplier_id = $id");
+$data = mysqli_fetch_assoc($cek);
+
+if ($data['jml'] > 0) {
+    echo "<script>
+            alert('Tidak dapat menghapus supplier karena masih digunakan di tabel barang!');
+            window.location='index.php';
+          </script>";
+} else {
+    mysqli_query($koneksi, "DELETE FROM supplier WHERE id=$id");
+    header("Location: index.php");
+}
 ?>
